+++ Runtime/Utils/Box.cs
@@ -2,7 +2,7 @@ namespace DT.UniStart {
   /// <summary>
   /// Explicit box. Useful for passing by reference.
   /// </summary>
-  public class Box<T> {
+  public class Box<T> : IGetSetValue<T> {
     public T Value;
 
     public Box(T value) {